split evm tests into multiple units
